Move scanner to services/dexes/sovryn
