+++ example.env
@@ -75,3 +75,5 @@ NEW_RELIC_LICENSE_KEY=
 
 # NewRelic enable
 NEW_RELIC_ENABLED=false
+
+SOVRYN_SUBGRAPH_API=
