+++ src/services/dexes/sovryn/SovrynGraphResponseSchema.ts
