+++ src/services/dexes/sovryn/SovrynPoolScanner.ts
@@ -1,6 +1,6 @@
 import {Validator} from 'jsonschema';
 
-import {GraphClientBase} from '../graph/GraphClient';
+import {GraphClientBase} from '../../graph/GraphClient';
 import {PoolRepositoryBase} from './SovrynPoolRepository.js';
 import {SovrynGraphPoolsResponseJSONSchema} from './SovrynGraphResponseSchema.js';
 import {liquidityPoolsQuery} from './SovrynGraphQueries.js';
